move VortexBufReader to own file
