+++ vortex-io/src/lib.rs
@@ -7,6 +7,7 @@
 //! This crate provides core traits for positioned and streaming IO, and via feature
 //! flags implements the core traits for several common async runtimes and backing stores.
 
+pub use buf::*;
 #[cfg(feature = "object_store")]
 pub use object_store::*;
 pub use read::*;
@@ -14,6 +15,7 @@ pub use read::*;
 pub use tokio::*;
 pub use write::*;
 
+mod buf;
 #[cfg(feature = "compio")]
 mod compio;
 #[cfg(feature = "object_store")]
